{"schema":"decodex/commit/1","summary":"Split MCP tool schemas","authority":"XY-1115"}
